Changed Logic on Dotfiles
Initial logic on dotfiles was that this lib should not support extension wrangling because of `.htaccess`. This further expanded to dotfiles with extension because traditionally *nix systems like mac OS & Linux, BSD etc hide dotfiles.

Revised logic is that this is an OS-specific problem and doesn't cross over the domain of this project. A Public `isDotFile` and `hasNoExtension` is provided, so really this doesn't matter.

Signed-off-by: Lewis Cowles <lewiscowles@me.com>