look into travis on 7.4 runtime (#12)
* look into travis on 7.4 runtime

* Allow failures for old runtimes

Any EOL runtime can accept failure, and forward compat is more important than backwards compat. This is not as the result of any changes but as an alternative to pinning PHPUnit for these builds. If you were to install an older PHPUnit, I believe these should work and an issue raised if not.

That said they will soon be out of scope for the repo